requirements.txt
mujoco
numpy
roboticstoolbox-python[collision]
open3d
scipy
transformations